src/pyproject.toml
[tool.poetry]
name = "profile"
version = "0.1.0"
description = "Create the resume"
authors = ["Saravanakumar Arunachalam <saravanastar@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.1.2"
docx2pdf = "^0.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
